+++ src/Command.php
@@ -4,11 +4,7 @@ declare(strict_types=1);
 
 namespace Yiisoft\Db\Mysql;
 
-use PDOException;
-use Throwable;
 use Yiisoft\Db\Driver\Pdo\AbstractPdoCommand;
-use Yiisoft\Db\Exception\ConvertException;
-use Yiisoft\Db\QueryBuilder\QueryBuilderInterface;
 
 /**
  * Implements a database command that can be executed with a PDO (PHP Data Object) database connection for MySQL,
@@ -22,11 +18,12 @@ final class Command extends AbstractPdoCommand
         $sql = $this->db->getQueryBuilder()->insert($table, $columns, $params);
         $this->setSql($sql)->bindValues($params);
 
+        $tableSchema = $this->db->getSchema()->getTableSchema($table);
+
         if (!$this->execute()) {
             return false;
         }
 
-        $tableSchema = $this->db->getSchema()->getTableSchema($table);
         $tablePrimaryKeys = $tableSchema?->getPrimaryKey() ?? [];
         $result = [];
 
@@ -50,47 +47,5 @@ final class Command extends AbstractPdoCommand
         SQL;
 
         return $this->setSql($sql)->queryColumn();
-    }
-
-    protected function getQueryBuilder(): QueryBuilderInterface
-    {
-        return $this->db->getQueryBuilder();
-    }
-
-    /**
-     * @psalm-suppress UnusedClosureParam
-     *
-     * @throws Throwable
-     */
-    protected function internalExecute(string|null $rawSql): void
-    {
-        $attempt = 0;
-
-        while (true) {
-            try {
-                if (
-                    ++$attempt === 1
-                    && $this->isolationLevel !== null
-                    && $this->db->getTransaction() === null
-                ) {
-                    $this->db->transaction(
-                        function () use ($rawSql): void {
-                            $this->internalExecute($rawSql);
-                        },
-                        $this->isolationLevel,
-                    );
-                } else {
-                    $this->pdoStatement?->execute();
-                }
-                break;
-            } catch (PDOException $e) {
-                $rawSql = $rawSql ?: $this->getRawSql();
-                $e = (new ConvertException($e, $rawSql))->run();
-
-                if ($this->retryHandler === null || !($this->retryHandler)($e, $attempt)) {
-                    throw $e;
-                }
-            }
-        }
     }
 }