+++ src/DQLQueryBuilder.php
@@ -4,9 +4,10 @@ declare(strict_types=1);
 
 namespace Yiisoft\Db\Mysql;
 
-use Yiisoft\Db\Expression\ExpressionBuilder;
+use Yiisoft\Db\Expression\Expression;
 use Yiisoft\Db\Expression\ExpressionInterface;
 use Yiisoft\Db\Expression\JsonExpression;
+use Yiisoft\Db\Mysql\Builder\ExpressionBuilder;
 use Yiisoft\Db\Mysql\Builder\JsonExpressionBuilder;
 use Yiisoft\Db\QueryBuilder\AbstractDQLQueryBuilder;
 
@@ -32,8 +33,8 @@ final class DQLQueryBuilder extends AbstractDQLQueryBuilder
             /**
              * Limit isn't optional in MySQL.
              *
-             * @link http://stackoverflow.com/a/271650/1106908
-             * @link http://dev.mysql.com/doc/refman/5.0/en/select.html#idm47619502796240
+             * @link https://stackoverflow.com/a/271650/1106908
+             * @link https://dev.mysql.com/doc/refman/5.0/en/select.html#idm47619502796240
              */
             $sql = 'LIMIT ' .
                 ($offset instanceof ExpressionInterface ? $this->buildExpression($offset) : (string)$offset) .
@@ -83,6 +84,7 @@ final class DQLQueryBuilder extends AbstractDQLQueryBuilder
             parent::defaultExpressionBuilders(),
             [
                 JsonExpression::class => JsonExpressionBuilder::class,
+                Expression::class => ExpressionBuilder::class,
             ]
         );
     }