+++ src/Schema.php
@@ -18,6 +18,7 @@ use Yiisoft\Db\Schema\Builder\ColumnInterface;
 use Yiisoft\Db\Schema\Column\ColumnSchemaInterface;
 use Yiisoft\Db\Schema\TableSchemaInterface;
 
+use function array_change_key_case;
 use function array_map;
 use function array_merge;
 use function array_values;
@@ -77,11 +78,11 @@ use function trim;
 final class Schema extends AbstractPdoSchema
 {
     /**
-     * @var array Mapping from physical column types (keys) to abstract column types (values).
+     * Mapping from physical column types (keys) to abstract column types (values).
      *
-     * @psalm-var string[]
+     * @var string[]
      */
-    private array $typeMap = [
+    private const TYPE_MAP = [
         'tinyint' => self::TYPE_TINYINT,
         'bit' => self::TYPE_INTEGER,
         'smallint' => self::TYPE_SMALLINT,
@@ -144,7 +145,7 @@ final class Schema extends AbstractPdoSchema
         $uniqueIndexes = [];
         $regexp = '/UNIQUE KEY\s+[`"](.+)[`"]\s*\(([`"].+[`"])+\)/mi';
 
-        if (preg_match_all($regexp, $sql, $matches, PREG_SET_ORDER)) {
+        if (preg_match_all($regexp, $sql, $matches, PREG_SET_ORDER) > 0) {
             foreach ($matches as $match) {
                 $indexName = $match[1];
                 $indexColumns = array_map('trim', preg_split('/[`"],[`"]/', trim($match[2], '`"')));
@@ -213,7 +214,7 @@ final class Schema extends AbstractPdoSchema
         /** @psalm-var ColumnInfoArray $info */
         foreach ($columns as $info) {
             /** @psalm-var ColumnInfoArray $info */
-            $info = $this->normalizeRowKeyCase($info, false);
+            $info = array_change_key_case($info);
 
             $info['extra_default_value'] = $columnsExtra[$info['field']] ?? '';
 
@@ -391,6 +392,8 @@ final class Schema extends AbstractPdoSchema
      * @param string $name The table name.
      *
      * @return array The cache key.
+     *
+     * @psalm-suppress DeprecatedMethod
      */
     protected function getCacheKey(string $name): array
     {
@@ -530,7 +533,7 @@ final class Schema extends AbstractPdoSchema
             }
         }
 
-        return $this->typeMap[$dbType] ?? self::TYPE_STRING;
+        return $this->TYPE_MAP[$dbType] ?? self::TYPE_STRING;
     }
 
     /**
@@ -558,7 +561,7 @@ final class Schema extends AbstractPdoSchema
             return new Expression('CURRENT_TIMESTAMP' . (!empty($matches[1]) ? '(' . $matches[1] . ')' : ''));
         }
 
-        if (!empty($column->getExtra()) && !empty($defaultValue)) {
+        if (!empty($defaultValue) && !empty($column->getExtra())) {
             return new Expression($defaultValue);
         }
 
@@ -658,7 +661,7 @@ final class Schema extends AbstractPdoSchema
         ])->queryAll();
 
         /** @psalm-var array[][] $constraints */
-        $constraints = $this->normalizeRowKeyCase($constraints, true);
+        $constraints = array_map('array_change_key_case', $constraints);
         $constraints = DbArrayHelper::index($constraints, null, ['type', 'name']);
 
         $result = [
@@ -773,7 +776,7 @@ final class Schema extends AbstractPdoSchema
         ])->queryAll();
 
         /** @psalm-var array[] $indexes */
-        $indexes = $this->normalizeRowKeyCase($indexes, true);
+        $indexes = array_map('array_change_key_case', $indexes);
         $indexes = DbArrayHelper::index($indexes, null, ['name']);
         $result = [];
 
@@ -898,7 +901,7 @@ final class Schema extends AbstractPdoSchema
         $result = [];
         $regexp = '/json_valid\([\`"](.+)[\`"]\s*\)/mi';
 
-        if (preg_match_all($regexp, $sql, $matches, PREG_SET_ORDER)) {
+        if (preg_match_all($regexp, $sql, $matches, PREG_SET_ORDER) > 0) {
             foreach ($matches as $match) {
                 $result[] = $match[1];
             }