+++ tests/QueryBuilderTest.php
@@ -4,7 +4,6 @@ declare(strict_types=1);
 
 namespace Yiisoft\Db\Mysql\Tests;
 
-use Generator;
 use Throwable;
 use Yiisoft\Db\Exception\Exception;
 use Yiisoft\Db\Exception\InvalidArgumentException;
@@ -143,9 +142,14 @@ final class QueryBuilderTest extends CommonQueryBuilderTest
      * @throws InvalidArgumentException
      * @throws NotSupportedException
      */
-    public function testBatchInsert(string $table, array $columns, iterable|Generator $rows, string $expected): void
-    {
-        parent::testBatchInsert($table, $columns, $rows, $expected);
+    public function testBatchInsert(
+        string $table,
+        array $columns,
+        iterable $rows,
+        string $expected,
+        array $expectedParams = [],
+    ): void {
+        parent::testBatchInsert($table, $columns, $rows, $expected, $expectedParams);
     }
 
     /**
@@ -512,7 +516,6 @@ final class QueryBuilderTest extends CommonQueryBuilderTest
     public function testRenameColumn(): void
     {
         $db = $this->getConnection();
-
         $qb = $db->getQueryBuilder();
 
         $this->assertSame(
@@ -520,6 +523,20 @@ final class QueryBuilderTest extends CommonQueryBuilderTest
             ALTER TABLE `alpha` CHANGE `string_identifier` `string_identifier_test` varchar(255) NOT NULL
             SQL,
             $qb->renameColumn('alpha', 'string_identifier', 'string_identifier_test'),
+        );
+
+        $this->assertSame(
+            <<<SQL
+            ALTER TABLE `alpha` CHANGE `non_exist_column` `new_column`
+            SQL,
+            $qb->renameColumn('alpha', 'non_exist_column', 'new_column'),
+        );
+
+        $this->assertSame(
+            <<<SQL
+            ALTER TABLE `non_exist_table` CHANGE `non_exist_column` `new_column`
+            SQL,
+            $qb->renameColumn('non_exist_table', 'non_exist_column', 'new_column'),
         );
     }
 
@@ -598,10 +615,11 @@ final class QueryBuilderTest extends CommonQueryBuilderTest
         string $table,
         array $columns,
         array|string $condition,
-        string $expectedSQL,
-        array $expectedParams
+        array $params,
+        string $expectedSql,
+        array $expectedParams,
     ): void {
-        parent::testUpdate($table, $columns, $condition, $expectedSQL, $expectedParams);
+        parent::testUpdate($table, $columns, $condition, $params, $expectedSql, $expectedParams);
     }
 
     /**
@@ -634,5 +652,11 @@ final class QueryBuilderTest extends CommonQueryBuilderTest
         array|bool $updateColumns
     ): void {
         parent::testUpsertExecute($table, $insertColumns, $updateColumns);
+    }
+
+    /** @dataProvider \Yiisoft\Db\Mysql\Tests\Provider\QueryBuilderProvider::selectScalar */
+    public function testSelectScalar(array|bool|float|int|string $columns, string $expected): void
+    {
+        parent::testSelectScalar($columns, $expected);
     }
 }