+++ tests/SchemaTest.php
@@ -541,7 +541,7 @@ final class SchemaTest extends CommonSchemaTest
         $this->expectException(NotSupportedException::class);
         $this->expectExceptionMessage('Only PDO connections are supported.');
 
-        $schema->refreshTableSchema('customer');
+        $schema->refresh();
     }
 
     public function testInsertDefaultValues()